feat: add conversation header and filter stories
