+++ ReactotronConfig.js
@@ -3,7 +3,7 @@ import Reactotron from 'reactotron-react-native';
 import { reactotronRedux } from 'reactotron-redux';
 
 const reactotron = Reactotron.useReactNative() // add all built-in react native plugins
-  // .use(reactotronRedux())
+  .use(reactotronRedux())
   .connect(); //Don't forget about me!
 
 export default reactotron;
