+++ src/screens/conversations/components/conversation-filters/index.ts
@@ -1,0 +1,1 @@
+export { ConversationFilterBar } from './ConversationFilterBar';
