+++ src/screens/conversations/components/conversation-header/index.ts
@@ -1,0 +1,1 @@
+export * from './ConversationHeaderContainer';
