+++ src/screens/conversations/components/index.ts
@@ -1,3 +1,5 @@
-export * from './ConversationHeader';
 export * from './conversation-item';
 export * from './conversation-sheet';
+export * from './conversation-header';
+export * from './conversation-filters';
+
