--- src/screens/conversations/index.ts
@@ -1,1 +1,0 @@
-export * from './ConversationList';
